make boldr smaller by moving middleware to its own class
