+++ src/server/server.dev.js
@@ -7,6 +7,7 @@ import { createServer } from 'http';
 import proxy from 'koa-proxy';
 import SocketIO from 'socket.io';
 import Boldr from './boldr';
+import BoldrMiddleware from './middleware';
 import projectConfig from 'config';
 import { logger as _log, normalizePort, onError } from './utils';
 import { handleRender } from './utils/renderReact';
@@ -15,13 +16,17 @@ dotenv.config();
 
 const debug = _debug('app:server:dev');
 const app = new Koa();
+// export server for easier testing.
 export const server = createServer(app.callback());
 const io = SocketIO(server);
+// Logging
 const log = _log(module);
+// Application constants
 const { SERVER_HOST, SERVER_PORT, WEBPACK_DEV_SERVER_PORT } = projectConfig;
 
-export function init() {
-  Boldr.init(app);
+async function init() {
+  await BoldrMiddleware.init(app);
+  await Boldr.initRoutes(app);
   /**
    * Loads the development specific functions
    * @param  {Boolean} __DEV__ Global variable for development environment
@@ -38,7 +43,7 @@ export function init() {
   // This is fired every time the server side receives a request
   app.use(handleRender);
 
-  server.listen(SERVER_PORT, () => {
+  await server.listen(SERVER_PORT, () => {
     log.debug(`Boldr server listening on ${SERVER_PORT} in ${process.env.NODE_ENV} node`);
   });
   server.on('error', onError);
